+++ resources/js/composables/mutations/theme/useThemeUpdateMutation.js
@@ -2,9 +2,7 @@ import { useMutation } from '@tanstack/vue-query'
 import axios from 'axios'
 
 const updateTheme = async (themeId, formData) => {
-  console.log(`themeId`, themeId)
-  console.log(`formData`, formData)
-  const { data } = await axios.patch(route(`api.themes.update`, themeId), {
+  const { data } = await axios.put(route(`api.themes.update`, themeId), {
     ...formData,
   })